Move `sessionSidebar` to be inside of `components` instead of `app/chat`
